api/requirements.txt
flask
flask_restful
pymongo
spacy
bcrypt
python-box
